@@ -14,20 +14,21 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@emotion/react": "^11.8.2",
+    "@emotion/styled": "^11.8.1",
     "@ericblade/quagga2": "^1.6.0",
-    "@material-ui/core": "^3.9.3",
-    "@material-ui/icons": "^3.0.2",
+    "@mui/material": "^5.5.3",
     "react": "^17.0.2",
     "react-countdown-now": "^2.1.0",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.0.0",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
     "react-textfit": "^1.1.0",
     "typeface-roboto": "^1.1.13"
   },
   "devDependencies": {
     "dotenv": "^16.0.0",
-    "eslint": "^6.8.0",
+    "eslint": "^8.12.0",
     "eslint-config-kokarn": "^4.0.4"
   },
   "engines": {
